% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateZcondonXY.R
\name{bindXYZpiI}
\alias{bindXYZpiI}
\title{Generates a dataframe with the grid (U), Y and Z1 and Z2.}
\usage{
bindXYZpiI(x1, x2, ...)
}
\arguments{
\item{length.out}{number of x coordinates for the grid}
}
\value{
a dataframe with two variables named x1 and x2.
}
\description{
Generates a dataframe with the grid (U), Y and Z1 and Z2.
}
\author{
DB
}
