% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure1.R
\name{figure1}
\alias{figure1}
\title{Generates the first figure of the paper}
\usage{
figure1(tikzexportfolder = NULL, pdfexportfolder = NULL,
  pngconvert = FALSE)
}
\arguments{
\item{tikzexportfolder}{a character string indicating a path where to save the figure in tikz format. If NULL (default), no tikz export is done.}
}
\value{
an object of classes "gg" and "ggplot": the first figure of the paper.
}
\description{
This function 
 - generates a regular grid of size 201 on the \eqn{[0,1]^2} square.
 - uses \code{InformativeSpatialSelection::generateYonsquaregridU} to generate the values of \eqn{Y}.
 - replicates the process 3 times 
 - generates a heatmap
}
\seealso{
InformativeSpatialSelection::generateYonsquaregridU
}
\author{
DB
}
