% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2.R
\name{figure2}
\alias{figure2}
\title{Generates the second figure of the paper}
\usage{
figure2(pdfexportfolder = NULL, pngconvert = FALSE)
}
\arguments{
\item{light}{a boolean (FALSE by default) indicating whether we should produce a light (with less pixels) version of the file.}
}
\value{
an object of classes "gg" and "ggplot": the first figure of the paper.
}
\description{
This function 
 - generates a regular grid of size 201 on the \eqn{[0,1]^2} square.
 - uses \code{InformativeSpatialSelection::generateYonsquaregridU} to generate the values of \eqn{Y}.
}
\seealso{
InformativeSpatialSelection::generateYonsquaregridU
}
\author{
DB
}
