% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2bis.R
\name{figure2bis}
\alias{figure2bis}
\title{Generates the first figure of the paper}
\usage{
figure2bis(tikzexportfolder = NULL, pdfexportfolder = NULL,
  light = FALSE, pngconvert = FALSE)
}
\arguments{
\item{tikzexportfolder}{a character string indicating a path where to save the figure in tikz format. If NULL (default), no tikz export is done.}

\item{light}{a boolean (FALSE by default) indicating whether we should produce a light (with less pixels) version of the file.}
}
\value{
an object of classes "gg" and "ggplot": the first figure of the paper.
}
\description{
This function 
 - generates a regular grid of size 201 on the \eqn{[0,1]^2} square.
 - uses \code{InformativeSpatialSelection::generateYonsquaregridU} to generate the values of \eqn{Y}.
}
\seealso{
InformativeSpatialSelection::generateYonsquaregridU
}
\author{
DB
}
