% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure3.R
\name{figure3}
\alias{figure3}
\title{Generates figure 3 of the paper}
\usage{
figure3(pdfexportfolder = NULL, light = FALSE, pngconvert = FALSE)
}
\arguments{
\item{light}{a boolean (FALSE by default) indicating whether we should produce a light (with less pixels) version of the file.}

\item{pngconvert}{a boolean (FALSE by default) indicating whether we should convet the pdf file into png. If yes, a png will be created}

\item{tikzexportfolder}{a character string indicating a path where to save the figure in tikz format. If NULL (default), no tikz export is done.}
}
\value{
an object of classes "gg" and "ggplot": the first figure of the paper.
}
\description{
Generates figure 3 of the paper
}
\seealso{
InformativeSpatialSelection::generateYonsquaregridU
}
\author{
FP
}
