% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure4.R
\name{figure4}
\alias{figure4}
\title{Generates figure 4 of the paper}
\usage{
figure4(pdfexportfolder = NULL, pngconvert = FALSE)
}
\arguments{
\item{pdfexportfolder}{a character string indicating a path where to save the figure in pdf format. If NULL (default), no pdf export is created.}

\item{pngconvert}{a boolean (FALSE by default) indicating whether we should convet the pdf file into png. If TRUE, a png will be created in the pdfexportfolder. If no pdf export folder is given, no png will be created}
}
\value{
an object of classes "gg" and "ggplot": the first figure of the paper.
}
\description{
Generates figure 4 of the paper
}
\seealso{
InformativeSpatialSelection::generateYonsquaregridU
}
\author{
FP
}
