% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateU.R
\name{generatesquaregridU}
\alias{generatesquaregridU}
\title{Simple function that generates U as a grid on the [0,1]^2 square.}
\usage{
generatesquaregridU(length.x1 = 100, length.x2 = length.x1,
  x1 = generatex_i(length.x1), x2 = generatex_i(length.x2))
}
\arguments{
\item{length.x1}{number of x coordinates for the grid (optional and ignored if x1 is given)}

\item{length.x2}{number of y coordinates for the grid (optional and ignored if x2 is given)}

\item{x1}{x coordinates for the grid (optional if length.x1 is given)}

\item{x2}{y coordinates for the grid (optional if length.x1 is given)}
}
\value{
a dataframe with two variables named x1 and x2.
}
\description{
Simple function that generates U as a grid on the [0,1]^2 square.
}
\author{
DB
}
