% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateU.R
\name{generatex_i}
\alias{generatex_i}
\title{Function that generates a sequence of points between 0 and 1.}
\usage{
generatex_i(length.out)
}
\arguments{
\item{length.out}{number of x coordinates for the grid}
}
\value{
a dataframe with two variables named x1 and x2.
}
\description{
Function that generates a sequence of points between 0 and 1.
}
\author{
DB
}
