% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateY.R
\name{generateYonsquaregridU}
\alias{generateYonsquaregridU}
\title{Generates a Gaussian spatial process \eqn{Y} with a Guassian semivariogram}
\usage{
generateYonsquaregridU(length.out = 100, x1 = generatex_i(length.out),
  x2 = x1, Variogram = "Gaussian", .scale = 10, mu = 10,
  .sigma = 1)
}
\arguments{
\item{x1}{a numerical vector.}

\item{x2}{a numerical vector.}

\item{Variogram}{(default \code{"Gaussian"}): type of variogram.}

\item{.scale}{a numerical value (Default 10)}

\item{mu}{a numerical value (Default 20)}

\item{.sigma}{a positive numerical value (Default 1)}

\item{length.x1}{number of x coordinates for the grid, default to 100, which implies a grid of 100 * 100.}

\item{length.x2}{number of x coordinates for the grid, default to 100, which implies a grid of 100 * 100.}
}
\value{
a vector.
}
\description{
This function is based on \code{RandomFields::RFsimulate}. It uses the option 
\code{model = RandomFields::RMgauss() + RandomFields::RMtrend(mean = 0)}
}
\examples{
x1<-generatex_i(20)
Y<-generateYonsquaregridU(x1=x1)
UY<-cbind(x1,x1,y=Y)
library(ggplot2)
ggplot(data.frame(generatesquaregridU(x1=x1,x2=x1),y= Y),aes(x1,x2)) + 
  ggplot2::geom_tile(ggplot2::aes(fill = y),size=2) + 
  ggplot2::xlab("")+
  ggplot2::ylab("")+
  ggplot2::scale_x_continuous(expand=c(0,0))+
  ggplot2::scale_y_continuous(expand=c(0,0))+
  ggplot2::theme(
axis.text.x=ggplot2::element_blank(),
    axis.ticks.x=ggplot2::element_blank(),
    axis.text.y=ggplot2::element_blank(),
    axis.ticks.y=ggplot2::element_blank(),
    panel.border = ggplot2::element_blank(),
    panel.grid.major = ggplot2::element_blank(), 
    panel.grid.minor = ggplot2::element_blank(),
    panel.background = ggplot2::element_blank())+
  ggplot2::scale_fill_gradient(low = "white",     high = "black")  +
  ggplot2::labs(fill = "Signal")
}
\author{
DB
}
