% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateZcondonXY.R
\name{generateZconditionnallyongridUY}
\alias{generateZconditionnallyongridUY}
\title{Generates a Gaussian spatial process with a Guassian semivariogram}
\usage{
generateZconditionnallyongridUY(length.out = 100,
  x1 = generatex_i(length.out), x2 = x1, y, alpha = 5,
  beta = c(0.5, 10), .scale = 10, .sigma = 1,
  Variogram = "Gaussian")
}
\arguments{
\item{length.out}{number of x coordinates for the grid}
}
\value{
a dataframe with two variables named x1 and x2.
}
\description{
Generates a Gaussian spatial process with a Guassian semivariogram
}
\author{
DB
}
