% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_Data.R
\name{Simulate_fig4}
\alias{Simulate_fig4}
\title{Generates data for figure 4 of the paper}
\usage{
figure4_data()
}
\value{
a list of objects necessary for figure3.
}
\description{
Data simulation for figure 4.
}
