% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_Data.R
\name{Simulate_UY}
\alias{Simulate_UY}
\title{Simulate all X,Y,Z,the data for the project}
\usage{
Simulate_UY()
}
\value{
a dataframe.
}
\description{
Code that generates table  from package.
Simulates U and different replications of Y and Z.
}
