% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_Data.R
\name{Simulate_XYZ}
\alias{Simulate_XYZ}
\title{Simulate all X,Y,Z,the data for the project}
\usage{
Simulate_XYZ()
}
\value{
a dataframe.
}
\description{
Simulates U and different replications of Y and Z.
}
