% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{UY}
\alias{UY}
\title{Simulation UY}
\format{A data frame with 121203 rows and 6 columns:
\describe{
  \item{.n}{identification number of gaussian random field}
  \item{x1}{first coordinate}
  \item{x2}{second coordinate}
  \item{y}{response variable}
  \item{Z}{proxy variable}
  \item{I}{indicator variable}
         }}
\usage{
UY
}
\description{
The dataset contains three simulated gaussian random fields, with
\eqn{\mu=20}, \eqn{\sigma=5}, 
and scale = 10.
}
\keyword{datasets}
