% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{XYZ}
\alias{XYZ}
\title{Simulation XYZ}
\format{A data frame with 2500 rows and 6 columns:
\describe{
  \item{x1}{first coordinate}
  \item{x2}{second coordinate}
  \item{y}{response variable}
  \item{z.z1}{proxy variable}
  \item{z.z2}{proxy variable}
  \item{z.z3}{proxy variable}
         }}
\usage{
XYZ
}
\description{
The dataset contains a simulated random gaussian field, with
\eqn{\mu=20}, \eqn{\sigma=5}, scale = 10, and corresponding proxy variables.
}
\keyword{datasets}
